--- Controls-demo/grid/EditInPlace/Base/Base.wml
@@ -1,1 +1,0 @@
-<Controls-demo.WIP byTask="https://online.sbis.ru/opendoc.html?guid=d39c25fd-3b40-45d3-9143-2bd868492f4a"/>
