--- Controls-demo/grid/EditInPlace/Base/Index.ts
@@ -1,19 +1,0 @@
-import {Control, TemplateFunction} from "UI/Base"
-import * as Template from "wml!Controls-demo/grid/EditInPlace/Base/Base"
-import {Memory} from "Types/source"
-import {getPorts} from "../../DemoHelpers/DataCatalog"
-
-import 'css!Controls-demo/Controls-demo'
-
-export default class extends Control {
-    protected _template: TemplateFunction = Template;
-    protected _viewSource: Memory;
-    protected _columns = getPorts().getColumns();
-
-    protected _beforeMount() {
-        this._viewSource = new Memory({
-            keyProperty: 'id',
-            data: getPorts().getData()
-        });
-    }
-}
