+++ Controls-demo/grid/EditInPlace/EditInPlace.wml
@@ -2,10 +2,10 @@
     <div class="controlsDemo__cell">
         <Router.router:Reference attr:class="controls-text-label_theme-{{_options.theme}}"
                                  state="/Controls-demo/app/:app"
-                                 app="{{ 'Controls-demo/grid/EditInPlace/Base/Index' }}">
-            <a href="{{content.href}}">Базовая настройка редактирования по месту</a>
+                                 app="{{ 'Controls-demo/grid/EditInPlace/EditingOnMounting/Index' }}">
+            <a href="{{content.href}}">Таблица с запущеным добавлением при построении</a>
         </Router.router:Reference>
-        <Controls-demo.grid.EditInPlace.Base.Index/>
+        <Controls-demo.grid.EditInPlace.EditingOnMounting.Index/>
     </div>
 
     <div class="controlsDemo__cell">
