+++ Controls/_treeGrid/TreeGrid/RowEditor.wml
@@ -1,5 +1,5 @@
 <ws:partial
-   template="{{itemData.resolveBaseItemTemplate()}}"
+   template="{{itemData.resolvers.baseItemTemplate()}}"
    scope="{{_options}}"
    colspanTemplate="{{itemData.isEditing ? content}}"
    colspan="{{itemData.isEditing}}"/>
